[DSPlayer] FilterCoreFactory use a stream index instead of a SStreamInfos struct
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/dsplayer@28723 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/DSPlayer/DSGraph.h
@@ -203,10 +203,10 @@ private:
       player_state = "";
       current_filter_state = State_Stopped;
     }
-    double timestamp;         // last time of update
+    LONGLONG timestamp;         // last time of update
 
-    double time;              // current playback time in millisec
-    double time_total;        // total playback time in millisec
+    LONGLONG time;              // current playback time in millisec
+    LONGLONG time_total;        // total playback time in millisec
     FILTER_STATE current_filter_state;
 
     std::string player_state;  // full player state
